cyberimpact_backend/requirements.txt
fastapi
uvicorn
pydantic
gitpython
google-generativeai
bandit
safety
python-docx
semgrep
njsscan
firebase-admin
pymongo
python-multipart
openpyxl
PyPDF2
